Update to use templates and move to main package
